+++ client/doublezerod/cmd/doublezerod/main.go
@@ -37,11 +37,12 @@ var (
 	routeConfigPath      = flag.String("route-config", "/var/lib/doublezerod/route-config.json", "path to route config file (unstable)")
 
 	// Route liveness configuration flags.
-	routeLivenessTxMin      = flag.Duration("route-liveness-tx-min", defaultRouteLivenessTxMin, "route liveness tx min")
-	routeLivenessRxMin      = flag.Duration("route-liveness-rx-min", defaultRouteLivenessRxMin, "route liveness rx min")
-	routeLivenessDetectMult = flag.Uint("route-liveness-detect-mult", defaultRouteLivenessDetectMult, "route liveness detect mult")
-	routeLivenessMinTxFloor = flag.Duration("route-liveness-min-tx-floor", defaultRouteLivenessMinTxFloor, "route liveness min tx floor")
-	routeLivenessMaxTxCeil  = flag.Duration("route-liveness-max-tx-ceil", defaultRouteLivenessMaxTxCeil, "route liveness max tx ceil")
+	routeLivenessTxMin       = flag.Duration("route-liveness-tx-min", defaultRouteLivenessTxMin, "route liveness tx min")
+	routeLivenessRxMin       = flag.Duration("route-liveness-rx-min", defaultRouteLivenessRxMin, "route liveness rx min")
+	routeLivenessDetectMult  = flag.Uint("route-liveness-detect-mult", defaultRouteLivenessDetectMult, "route liveness detect mult")
+	routeLivenessMinTxFloor  = flag.Duration("route-liveness-min-tx-floor", defaultRouteLivenessMinTxFloor, "route liveness min tx floor")
+	routeLivenessMaxTxCeil   = flag.Duration("route-liveness-max-tx-ceil", defaultRouteLivenessMaxTxCeil, "route liveness max tx ceil")
+	routeLivenessPeerMetrics = flag.Bool("route-liveness-peer-metrics", false, "enables per peer metrics for route liveness (high cardinality)")
 
 	// TODO(snormore): These flags are temporary for initial rollout testing.
 	// They will be superceded by a single `route-liveness-enable` flag, where false means
@@ -155,11 +156,12 @@ func main() {
 			// The manager only knows about passive mode, with the negation of it being active mode.
 			PassiveMode: !*routeLivenessEnableActive,
 
-			TxMin:      *routeLivenessTxMin,
-			RxMin:      *routeLivenessRxMin,
-			DetectMult: uint8(*routeLivenessDetectMult),
-			MinTxFloor: *routeLivenessMinTxFloor,
-			MaxTxCeil:  *routeLivenessMaxTxCeil,
+			TxMin:       *routeLivenessTxMin,
+			RxMin:       *routeLivenessRxMin,
+			DetectMult:  uint8(*routeLivenessDetectMult),
+			MinTxFloor:  *routeLivenessMinTxFloor,
+			MaxTxCeil:   *routeLivenessMaxTxCeil,
+			PeerMetrics: *routeLivenessPeerMetrics,
 		}
 	}